% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blast.R
\name{blastp}
\alias{blastp}
\title{Blast query sequences against reference sequences}
\usage{
blastp(ref, query, run_id, cmd_makeblastdb = "makeblastdb",
  cmd_blastp = "blastp", blastp_num_threads = "16", verbose = F)
}
\arguments{
\item{ref}{list of vector of chars where each element is a sequence object of class SeqFastadna or SeqFastaAA, e.g. what is returned by seqinr::read.fasta}

\item{query}{list of vector of chars where each element is a sequence object of class SeqFastadna or SeqFastaAA, e.g. what is returned by seqinr::read.fasta}

\item{run_id}{character string used to name temporary files etc.}

\item{cmd_makeblastdb}{the command used to make the reference database e.g. makeblastdb from the blast+ package}

\item{cmd_blastp}{the command used to blast the queries against the reference sequences e.g. blastp
blastp_num_threads the number of threads to use while running blast}

\item{verbose}{write out additional information while executing.}
}
\description{
requires makeblastdb and blastp from blast+ to be on the path
}
