% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clustalo.R
\name{clustalo}
\alias{clustalo}
\title{Run clustal omega}
\usage{
clustalo(fasta, output_fname, run_script_fname = NULL, ncores = 18,
  max_mem = 10000, output_fmt = "phylip", seqtype = "protein",
  force = F)
}
\description{
Run clustal omega
}

