% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phyml.R
\name{phyml}
\alias{phyml}
\title{Run phyml}
\usage{
phyml(phylip, run_id = NULL, run_script_fname = NULL,
  data_type = "aa", model = "LG", search = "NNI", ncores = 18,
  max_mem = 10000, output_fmt = "phylip", seqtype = "protein")
}
\description{
Run phyml
}
