% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/naming.R
\name{rename_tree}
\alias{rename_tree}
\title{Rename a tree read in by ape}
\usage{
rename_tree(tree, name_map, old_names_col = "phylip_name",
  new_names_col = "tree_name")
}
\description{
Rename a tree read in by ape
}

