% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sequences_to_tree.R
\name{sequences_to_tree}
\alias{sequences_to_tree}
\title{This is the full pipeline taking path to a fasta sequence or a
seqinr AA file and generating a newick tree}
\usage{
sequences_to_tree(sequences, output_path, run_id, user_agent_arg,
  sequence_name_to_uniprot_entry = hmmer_fasta_to_uniprot_entry,
  verbose = T)
}
\description{
This is the full pipeline taking path to a fasta sequence or a
seqinr AA file and generating a newick tree
}

